+++ src/jtl/Connector/Presta/Controller/ProductStockLevel.php
@@ -31,6 +31,7 @@ class ProductStockLevel extends BaseController
         if (!empty($id)) {
             if (strpos($id, '_') === false) {
                 \StockAvailable::setQuantity($id, null, $data->getStockLevel());
+                \StockAvailable::setProductOutOfStock($id, $model->out_of_stock == 1 ? true : false);
             } else {
                 list($productId, $combiId) = explode('_',$id);